Create create-workspace.bat
feat: Add create-workspace.bat script

This commit adds a batch script for creating a new workspace in Visual Studio Code. 
The script automatically detects the parent folder name as the workspace name and 
adds all folders within the directory to the workspace. It can be run by saving it as 
a `.bat` file in the root directory of the workspace and double-clicking on it.

The script also includes a `-h` option for displaying a help message.

This commit adds a new feature to the project.