feat: Add script for cloning multiple repositories
The batch script allows users to clone multiple git repositories using a project name to locate a dependency file with repository URLs. The dependency file should be located in a "repositories" folder with the same name as the project folder. The script also includes a help option, which can be accessed through the "-h" argument.